refactor: extracted game view as a separate class, made the GUI responsive
